refactor(tests): extract driver tests into part files
Moved the bulk of the integration tests from `driver.rs` into a new `driver_part_01_tests.rs` file, leaving only the `include!` directives that pull in the two part files. This keeps the driver module as a thin dispatcher and makes each test file shorter and easier to navigate.

Auto-committed-on: dragonfly
